вынос в JS из PHP файлов

+++ components/refbooks.import/component.php
@@ -17,37 +17,51 @@ if(!empty($_FILES['import_file']) && !empty($arParams['FILE_FIELDS'])){
 	global $DB;
 	global $XMLMapping;
 	$arDataMapping = $XMLMapping->getXMLDataMapping($arParams['MAPPING_CODE']);
-	$DB->query("TRUNCATE TABLE {$arDataMapping['MYSQL_TABLE_NAME']}");
-	$sql = "INSERT INTO {$arDataMapping['MYSQL_TABLE_NAME']} (##STRFIELDS##) VALUES (##STRVALUES##)";
+	if ($arParams['IS_UPDATE'] == 'Y'){
+		$sql = "UPDATE {$arDataMapping['MYSQL_TABLE_NAME']} SET ##STRVALUES## WHERE #MAIN_FIELD#=#MAIN_FIELD_VALUE#";
+	}else{
+		$DB->query("TRUNCATE TABLE {$arDataMapping['MYSQL_TABLE_NAME']}");
+		$sql = "INSERT INTO {$arDataMapping['MYSQL_TABLE_NAME']} (##STRFIELDS##) VALUES (##STRVALUES##)";
+	}
 	$arFields = array();
 	$arValues = array();
 	$counter = 0;
-	
-	$reader = new XMLReader(); 
-  $reader->open($_FILES["import_file"]["tmp_name"]); 
-	while ($reader->read()) { 
-	 switch ($reader->nodeType) { 
-	   case (XMLREADER::ELEMENT): 
+
+	$reader = new XMLReader();
+  $reader->open($_FILES["import_file"]["tmp_name"]);
+	while ($reader->read()) {
+	 switch ($reader->nodeType) {
+	   case (XMLREADER::ELEMENT):
 	   //var_dump($reader->name);
-				$fieldName = $reader->name; 
-				if ( in_array($fieldName , $arParams['FILE_FIELDS']) &&  array_key_exists($fieldName ,$arDataMapping)){ 
+				$fieldName = $reader->name;
+				if ( in_array($fieldName , $arParams['FILE_FIELDS']) &&  array_key_exists($fieldName ,$arDataMapping)){
 					if(!in_array($arDataMapping[$fieldName], $arFields)){
-						$arFields[] = $arDataMapping[$fieldName]; 
 						$reader->read();
-						$fieldValue = prepareRefBooksValue($reader->value); 
+						$fieldValue = prepareRefBooksValue($reader->value);
 						if(!empty($arDataMapping['USER_FUNCTION'])){
 							$arDataMapping['USER_FUNCTION']($fieldName, &$fieldValue);
 						}
-						$arValues[] = $fieldValue;
+						if ($fieldName == $arParams['MAIN_FIELD'] ){
+							$MAIN_FIELD = $arDataMapping[$arParams['MAIN_FIELD'] ];
+							$MAIN_FIELD_VALUE = $fieldValue;
+						}else{
+							$arFields[] = $arDataMapping[$fieldName];
+							$arValues[] = $fieldValue;
+						}
 					}
 				}elseif ($reader->name == $arParams['PARENT_TAG'] ){
-//					var_dump(count($arValues));
-//					var_dump(count($arFields));
 					if(count($arValues) > 0 && count($arValues) == count($arFields)){
 						$strFields = implode(", ", $arFields);
 						$strValues = "'".implode("', '", $arValues)."'";
 
-						$query = str_replace(array('##STRFIELDS##', '##STRVALUES##'), array($strFields, $strValues), $sql);
+						if ($arParams['IS_UPDATE'] == 'Y'){
+							$strValues = '#FIELD#="'. implode('", #FIELD#="', $arValues) .'"';
+							$strValues = str_replace(array('#FIELD#'), $arFields, $strValues);
+							$query = str_replace(array('##STRVALUES##', '#MAIN_FIELD#', '#MAIN_FIELD_VALUE#'), array($strValues, $MAIN_FIELD, $MAIN_FIELD_VALUE), $sql);
+							echo ($query).'<br/>';
+						}else{
+							$query = str_replace(array('##STRFIELDS##', '##STRVALUES##'), array($strFields, $strValues), $sql);
+						}
 						echo '<p>Row: '.++$counter;
 						$res = $DB->setRecord($query);
 						if( intval($res) < 1 ){
@@ -59,14 +73,16 @@ if(!empty($_FILES['import_file']) && !empty($arParams['FILE_FIELDS'])){
 					}
 					$arFields = array();
 					$arValues = array();
+					$MAIN_FIELD = '';
+					$MAIN_FIELD_VALUE = '';
 				}
-			break; 
-	
+			break;
+
 		}
 	}
 	if($arParams['DEBUG']){
 		echo '<pre>'. print_r($arError).'</pre>';
 	}
-	
+
 }
 ?>
